fix type missmatch

+++ NEK/Nikon.hpp
@@ -19,8 +19,8 @@ namespace nek {
 
 
 		mtp::MtpDevicePropDescDS GetDevicePropDesc(uint32_t devicePropCode);
-		mtp::MtpDatatypeVariant GetDevicePropValue(uint16_t devicePropCode);
-		void SetDevicePropValue(uint16_t devicePropCode, mtp::MtpDatatypeVariant data);
+		mtp::MtpDatatypeVariant GetDevicePropValue(uint32_t devicePropCode);
+		void SetDevicePropValue(uint32_t devicePropCode, mtp::MtpDatatypeVariant data);
 
 
 	private:
